+++ crates/miden-lib/src/account/auth/mod.rs
@@ -3,7 +3,7 @@ use miden_objects::{
     crypto::dsa::rpo_falcon512::PublicKey,
 };
 
-use crate::accounts::components::rpo_falcon_512_library;
+use crate::account::components::rpo_falcon_512_library;
 
 /// An [`AccountComponent`] implementing the RpoFalcon512 signature scheme for authentication of
 /// transactions.
