+++ crates/miden-lib/src/account/components/mod.rs
