+++ crates/miden-lib/src/note/mod.rs
