+++ crates/miden-lib/src/note/scripts.rs
